requirements.txt
asgiref==3.8.1
attrs==25.1.0
click==8.1.8
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==24.3
django-split-settings==1.3.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
factory_boy==3.3.3
Faker==36.1.0
h11==0.14.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
psycopg2-binary==2.9.10
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==4.1.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
uvicorn==0.34.0
